+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Last changes:
 
 
+Version `2.5 (18 Apr'20)`
+
+  - text and entry-like widgets (cbx ent fco spx) got their popup menu
+  - corrected popup menu for text widget of dialogs
+  - method setTextContents to fill contents of disabled text widgets
+  - choosers' names can be titled upper-case
+  - little clean-up
+
+
 Version `2.4 (15 Apr'20)`
 
-- v2.4
   - selection attributes for listboxes, to fix a listbox curselection issue
     as hinted by
       Donal K. Fellow (https://stackoverflow.com, the question